Fix: validaçoes no crud de alunos

+++ src/modules/pessoas/use-cases/delete-aluno/delete-aluno-use-case.ts
@@ -1,6 +1,6 @@
 import { AlunoRepository } from '@modules/pessoas/infra/repositories/aluno-repository';
 import { IAlunoRepository } from '@modules/pessoas/repositories/i-aluno-repository';
-import { HttpResponse, serverError } from '@shared/helpers';
+import { HttpResponse, notFound, serverError } from '@shared/helpers';
 import { inject, injectable } from "tsyringe";
 
 
@@ -13,6 +13,12 @@ class DeleteAlunoUseCase {
 
   async execute(id: string): Promise<HttpResponse> {
     try {
+
+      const alunoExists = await this.alunoRepository.get(id)
+
+      if(alunoExists.data === null) {
+        return notFound("Aluno não encontrado.")
+      }
 
       const result = await this.alunoRepository.delete(id)
